First DEN Service implementation within OScar
This commit introduces a first implementation of the DEN service within OScar (and a new related option --enable-DENMs-tx), with the following limitations:
(1) Reception and storage of events in the LDM is not yet supported, except for a simple print for logging the received DENMs; therefore isNegation has not been tested yet
(2) DCC and MetricSupervisor are not yet integrated
(3) The JSON-over-TCP interface for triggering DENMs is still a draft and shall be extended to support more fields and DENM types
(4) Destination areas for GBC packets can only be circular
(5) V2X security is not supported yet (DENMs will be sent without security even when --enable-security is specified)
(6) The code still needs to be polished and there are several TODOs that shall be tackled

This commit also fixed a series of compilation warnings.

+++ src/basicSensorReader.cpp
@@ -268,7 +268,7 @@ bool BasicSensorReader::startReader() {
     // Prepare a CAN ID filter
     // Filter only for the messages of interest
     // The data of interest should be located inside the messages matching the regex used to read the CAN database (default CAN message names: "Video_Object_0x_B")
-    for(int can_id_idx=0;can_id_idx<m_can_db_id_info.size();can_id_idx++) {
+    for(int can_id_idx=0;can_id_idx<static_cast<int>(m_can_db_id_info.size());can_id_idx++) {
         canfilter[can_id_idx].can_id=m_can_db_id_info[can_id_idx];
         canfilter[can_id_idx].can_mask=CAN_SFF_MASK;
     }
